requirements.txt
Flask==2.2.3
geographiclib==2.0
geopy==2.3.0
numpy==1.21.5
opencv_python==4.7.0.72
paho_mqtt==1.6.1
